% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate.smet}
\alias{aggregate.smet}
\alias{aggregate}
\title{aggregate}
\usage{
\method{aggregate}{smet}(x, date.field = "timestamp", FUN = mean,
  INDEX = by[1], by = c("hourly", "daily", "monthly", "yearly"),
  past = TRUE, ...)
}
\arguments{
\item{x}{a \code{\link{smet-class}} object}

\item{...}{further arguments}
}
\description{
aggregate
}
\examples{
smet <- as.smet(system.file("examples/T0179.smet",package="RSMET"))

out <- aggregate(smet)



}
