% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.meteoioini.R
\docType{methods}
\name{as.meteoioini}
\alias{as.meteoioini}
\alias{as.meteoioini,character-method}
\alias{as.meteoioini,list-method}
\alias{as.meteoioini,meteoioini-method}
\title{Coerces an object to a \code{meteoioini-class}  object}
\usage{
as.meteoioini(object, ...)

as.meteoioini(object, ...)

\S4method{as.meteoioini}{character}(object, ...)

\S4method{as.meteoioini}{list}(object, ...)

\S4method{as.meteoioini}{meteoioini}(object, ...)
}
\arguments{
\item{object}{the object to be coerced}

\item{...}{further arguments}
}
\description{
Coerces an object to a \code{meteoioini-class}  object








}
\examples{
ini <- as.meteoioini("test")
}

