% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/collapse.smet.R
\name{collapse.smet}
\alias{collapse.smet}
\title{Coerces a \code{smet-class} object to a data frame}
\usage{
collapse.smet(x, y, headers = NULL, date.field = "timestamp")
}
\arguments{
\item{x}{first \code{\link{smet-class}} object}

\item{y}{second \code{\link{smet-class}} object}

\item{headers}{header attributes to  check before collapsing}
}
\value{
a collapsed \code{\link{smet-class}} object
}
\description{
Coerces a \code{smet-class} object to a data frame
}
\examples{
smet1 <- as.smet(system.file("examples/zugspitze1.smet",package="RSMET"))
smet2 <- as.smet(system.file("examples/zugspitze2.smet",package="RSMET"))

newsmet <- collapse.smet(x=smet1,y=smet2)

newsmet_ <- collapse.smet(x=smet2,y=smet1)
}

