% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zbracket[.smet.R
\name{[.smet}
\alias{Extract}
\alias{[}
\alias{[.smet}
\title{\code{[} S3 method for \code{\link{smet-class}} object}
\usage{
\method{[}{smet}(x, rows = "all", fields = "all", file = NA,
  date.field = "timestamp", ...)
}
\arguments{
\item{x}{a \code{\link{smet-class}} object}

\item{rows}{rows of the \code{x} data which will be returned.}

\item{fields}{column number or fields of \code{x} which will be returned.}

\item{file}{file of the new returened \code{\link{smet-class}}. Default is \code{NA}.}

\item{date.field}{field name of the Date and Time variable. Default is \code{"timestamp"}. If it is not \code{NA}, it is always returned.}

\item{...}{further argument for \code{\link{[}} method}
}
\description{
\code{[} S3 method for \code{\link{smet-class}} object
}
\examples{

x <- smet(system.file('examples/PIEM001114.smet',package="RSMET"))
x[,1:3]


}
\author{
Emanuele Cordano
}

