% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zbracket[.replacement.smet.R
\name{[<-.smet}
\alias{[<-,extract_replacemethod}
\alias{[<-.smet}
\title{\code{'[<-'} S3  Replacement method for \code{smet-class} object}
\usage{
\method{[}{smet}(x, rows = "all", fields = "all",
  date.field = "timestamp", offset = 0, multiplier = 1, nodata = -9999,
  ...) <- value
}
\arguments{
\item{x}{a \code{\link{smet-class}} object}

\item{rows}{rows of the \code{x} data which will be replaced.}

\item{fields}{column number or fields of \code{x} which will be replaced.}

\item{date.field}{field name of the Date and Time variable. Default is \code{"timestamp"}. If it is not \code{NA}, it is always returned.}

\item{offset}{scalar value of the unit multiplier for each new \code{field} name which is not previously present.}

\item{multiplier}{scalar value of the unit multiplier for each new \code{field} name which is not previously present.}

\item{nodata}{numeric value used to replce \code{NA}. It is used if no-data value is not specified in the header of \code{x}.}

\item{...}{further argument for \code{\link{[}} method}

\item{value}{a object to be replaced}
}
\value{
The "replaced" \code{\link{smet-class}} object.
}
\description{
\code{'[<-'} S3  Replacement method for \code{smet-class} object
}
\note{
In case \code{fields} is a character vector, the elements whose names is in \code{value} is replaced.
}
\examples{
rm(list=ls())


 ### SMET Rifugio Vaccarone
}
\author{
Emanuele Cordano
}

