% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\docType{methods}
\name{fields}
\alias{fields}
\alias{fields,smet-method}
\title{Extract the fields  of  a \code{smet-class}  object}
\usage{
fields(object, ...)

fields(object, ...)

\S4method{fields}{smet}(object, ...)
}
\arguments{
\item{object}{the object}

\item{...}{further arguments}
}
\description{
The method \code{as.smet} coerces an object or a charachter string to a SMET. If the object is

fields

fields
}
\examples{


df <- data.frame(a=1:6,c=1:6)
fields(df)

x <- smet(system.file('examples/PIEM001114.smet',package="RSMET"))
fields(x)





}

