% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_smet_ini.R
\docType{class}
\name{meteoioini-class}
\alias{meteoioini-class}
\title{meteoioini-class}
\description{
Class \code{meteoioini}
}
\details{
This class represents the MeteoIO ini format file for weather station data 
 \describe{
    \item{\code{General}:}{General envirment (list)}

\item{\code{Input}:}{Input envirment (list)}

\item{\code{Output}:}{Output envirment (list)}

\item{\code{Filters}:}{Filters envirment (list)}

\item{\code{Interpolation2D}:}{Interpolation2D envirment (list)}

\item{\code{file}:}{full name of the SMET file. If it is missing, it is \code{NA}.}

}

Detailed information about SMET format is reminded to \url{http://models.slf.ch/docserver/meteoio/SMET_specifications.pdf}.
}
\note{
A \code{meteoioini-class} object can be created by a MeteoIO  file
}
\examples{

showClass("meteoioini")
as.meteoioini("test")

}
\author{
Emanuele Cordano
}
\references{
\url{http://models.slf.ch/docserver/meteoio/html/configuration.html}
}
\seealso{
\code{\link{meteoioini}},\code{\link{print.meteoioini}}
}
\keyword{classes}

