% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ini-function.R
\name{meteoioini}
\alias{meteoioini}
\title{Function to read a  a Meteo IO *.ini file}
\usage{
meteoioini(file = NULL, comment = "#", ...)
}
\arguments{
\item{file}{Meteo IO INI file name}

\item{comment}{character symbol used to comment lines or part of lines. Default is \code{"#"}. See \code{\link{str_locate}}}

\item{...}{further arguments}
}
\value{
a \code{\link{meteoioini-class}} object
}
\description{
Function to read a  a Meteo IO *.ini file
}
\examples{

file <- system.file("examples/io.ini",package="RSMET")
ini <- meteoioini(file)

}
\seealso{
\code{\link{meteoioini-class}},\code{\link{print.meteoioini}}
}

