% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.meteioini.R
\docType{methods}
\name{print.meteoioini}
\alias{print.meteoioini}
\alias{print_meteioini}
\alias{show,meteoioini-method}
\alias{show_meteioini}
\title{"Print" and "Show" method for "meteoioini-class" object}
\usage{
\method{print}{meteoioini}(x, file = NULL, print.all = !identical(file,
  NULL), max.records = 200, ...)

\S4method{show}{meteoioini}(object)
}
\arguments{
\item{x, object}{a \code{meoioini-class} object}

\item{file}{filename where to print the \code{meteoioini-class} object \code{x} as an ASCII file. If it is equal to \code{"internal"}, the filename is taken from \code{slot(x,"file")} (internally defined in \code{x}).}

\item{print.all}{logical value. If it is \code{FALSE} exceeding lines are omitted for a better human readibility.}

\item{max.records}{maximum printable number of records. Default is 200 and it is activated only if \code{print.all==FALSE}. ,}

\item{...}{further argumnents for \code{writeLines}}
}
\description{
"Print" and "Show" method for "meteoioini-class" object
}
\examples{

file <- system.file("examples/io.ini",package="RSMET")
ini <- meteoioini(file)
print(ini)


}
\seealso{
\code{\link{smet-class}}, \code{\link{smet}},\code{\link{writeLines}}
}

