% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/putsmatintoini.R
\name{putSmetIntoIni}
\alias{putSmetIntoIni}
\title{Put SMET object file into a \code{meteoini-class}}
\usage{
putSmetIntoIni(..., smetlist = NULL, ini = as.meteoioini("test"),
  smetwdir = system.file("temp", package = "RSMET"), force.smetdir = TRUE,
  new.ini.file = NA)
}
\arguments{
\item{...}{\code{\link{smet-class}} objects}

\item{smetlist}{(alternative) list of \code{\link{smet-class}} objects}

\item{ini}{a \code{\lik{meteoioini-class}} object}

\item{smetwdir}{.....}
}
\description{
Put SMET object file into a \code{meteoini-class}
}
\examples{
##  NOT YET IMPLEMENTED
}

