% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.R
\name{range.smet}
\alias{range}
\alias{range.smet}
\title{'range' method for a \code{\link{smet-class}} object}
\usage{
\method{range}{smet}(x, ..., field = "timestamp")
}
\arguments{
\item{x}{a \code{\link{smet-class}} object}

\item{...}{further arguments}

\item{field}{field (character string) used for range. Default is \code{"timestamp"}.}
}
\description{
'range' method for a \code{\link{smet-class}} object
}

