% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units_multiplier.R
\name{units_multiplier}
\alias{units_multiplier}
\title{Unit Multiplier of a \code{\link{smet}} object.}
\usage{
units_multiplier(x)
}
\arguments{
\item{x}{a \code{\link{smet-class}} object}
}
\description{
Unit Multiplier of a \code{\link{smet}} object.
}
\examples{

x <- smet(system.file('examples/PIEM001114.smet',package="RSMET"))
units_multiplier(x)

}

