% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units_offset.R
\name{units_offset}
\alias{units_offset}
\title{Unit Offsets of a \code{\link{smet}} object.}
\usage{
units_offset(x)
}
\arguments{
\item{x}{a \code{\link{smet-class}} object}
}
\description{
Unit Offsets of a \code{\link{smet}} object.
}
\examples{

x <- smet(system.file('examples/PIEM001114.smet',package="RSMET"))
units_offset(x)

}

