% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.R
\name{as.data.frame.smet}
\alias{as.data.frame}
\alias{as.data.frame.smet}
\title{Coerces a \code{smet-class} object to a data frame}
\usage{
\method{as.data.frame}{smet}(x, ..., date.field = "timestamp",
  add.header = FALSE, use.melt = FALSE)
}
\arguments{
\item{x}{a \code{smet-class} object}

\item{...}{further arguments}

\item{date.field}{field neme used for date and time. Default is \code{"timestamp"}.}

\item{add.header}{logical If \code{TRUE} it adds header scalar attributes as fields of the data frame. Default is \code{FALSE}.}

\item{use.melt}{logical. Default is \code{FALSE}. If it is \code{TRUE}, the result is melted with \code{\link{melt}}.}
}
\description{
Coerces a \code{smet-class} object to a data frame
}
\seealso{
\code{\link{smet-class}}, \code{\link{smet}},\code{\link{as.data.frame}}
}

